backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
aiomysql==0.2.0
pydantic==2.7.1
python-dotenv==1.0.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
paho-mqtt==1.6.1
scikit-learn==1.4.2
numpy==1.26.4
websockets==12.0
